Updated to libcouchbase 3.0.0-beta.2
This change also makes major changes to the error handling
and CAS semantics inside the SDK to reflect changes that
were made in libcouchbase which are incompatible with the
previous implementation.

Reviewed-on: http://review.couchbase.org/119873
Tested-by: Brett Lawson <brett19@gmail.com>

+++ deps/lcb/example/views/views-example.cc
@@ -50,7 +50,7 @@ static void viewCallback(lcb_INSTANCE *, int, const lcb_RESPVIEW *rv)
         rc = lcb_respget_status(doc);
         uint64_t cas;
         lcb_respget_cas(doc, &cas);
-        printf("   Document for response. RC=0x%X. CAS=0x%llx\n", rc, cas);
+        printf("   Document for response. RC=0x%X. CAS=0x%llx\n", rc, (long long)cas);
     }
 
     cbCounter++;
